Updated hdf5 to parquet converted to use pyarrow.
The first iteration of the parquet converter used both pandas and
hugging face datasets. Both of these use pyarrow under the hood. This
patch switches to using pyarrow only because it allows for a more
precise schema specification including support non-nullable types
(huggingface datasets only has nullable types). Arrow is closely
integrated with parquet and we therefore get the arrow schema reflected
in the parquet format.

This patch also changes the naming of the "embedding" column to "vector"
as it is more general.