Refactoring to use a dataset interface.
In preparation for the parquet dataset format this patch adds a dataset
interface to the benchmarker and provides an implementation of the
dataset interface for the existing HDF5 format.

This patch should not change the behavior of the benchmarker. Since we
do not have a testing suite it is possible that this patch has
introduced errors, but it should be a pure refactoring.