Updated Weaviate dependency for parquet support.
Changed conversion script back to using "embedding" naming.

Fixed a crashing bug for multivector that was introduced with the new
dataset interface. It happened when attempting to read the
dimensionality from the HDF5 file. For multivectors we pass the
dimension ourselves.

Ran local benchmarks comparing HDF5 and parquet. Import times, QPS, and
recall all look near-identical both for HNSW and the flat index. For the
flat index I see import times of approximately 3 seconds for
dbpedia-100k and 12 seconds for SIFT-1m. Parquet seems slightly faster
than HDF5, but it could be due to random variation or the order of
experiments.

Multivector (still from HDF5, not yet supported with parquet) and RQ
compression has been tested manually and matches expected behavior.